+++ src/main/scala/scala/tools/refactoring/implementations/oimports/TreeToolbox.scala
@@ -1,10 +1,6 @@
 package scala.tools.refactoring
 package implementations.oimports
 
-import scala.annotation.tailrec
-import scala.reflect.internal.util.RangePosition
-import scala.reflect.internal.util.SourceFile
-import scala.tools.nsc.ast.parser.Scanners
 import scala.tools.nsc.Global
 
 class TreeToolbox[G <: Global](val global: G) {
@@ -51,31 +47,6 @@ class TreeToolbox[G <: Global](val global: G) {
           case imp if !ancestorsImports.contains(kid.printImport(imp)) => imp
         })
       }
-    }
-  }
-
-  class TreeToolboxScanners extends {
-    val global = TreeToolbox.this.global
-  } with Scanners {
-    class CommentScanner(source: SourceFile) extends SourceFileScanner(source) { self: SourceFileScanner =>
-      private val comments_ = mutable.ListBuffer[RangePosition]()
-      override def skipComment(): Boolean = {
-        val start = this.offset
-        val result = super.skipComment()
-        if (result) {
-          comments_ += new RangePosition(source, start, start, this.charOffset)
-        }
-        result
-      }
-
-      def scan(): Unit = {
-        init()
-        import scala.tools.nsc.ast.parser.Tokens.EOF
-        @tailrec def scan(): Unit = if (token == EOF) () else { nextToken(); scan() }
-        scan()
-      }
-
-      def comments = comments_.toList
     }
   }
 }
